+++ src/markitdown/converters/__init__.py
@@ -8,6 +8,8 @@ from ._html_converter import HtmlConverter
 from ._rss_converter import RssConverter
 from ._wikipedia_converter import WikipediaConverter
 from ._youtube_converter import YouTubeConverter
+from ._ipynb_converter import IpynbConverter
+from ._bing_serp_converter import BingSerpConverter
 
 __all__ = [
     "DocumentConverter",
@@ -17,4 +19,6 @@ __all__ = [
     "RssConverter",
     "WikipediaConverter",
     "YouTubeConverter",
+    "IpynbConverter",
+    "BingSerpConverter",
 ]
